* chore: move from session source folder to connection.
The session folder contained all the source for talking
to the compute server/service. Naming it connection makes
more sense than session.
