+++ client/src/connection/rest/api/compute.ts
