+++ client/src/connection/rest/api/configuration.ts
