+++ client/src/connection/rest/auth.ts
